refactor(frontend): thermo-nuclear quality pass for copilot, explore, and mega-panels
Address structural findings from the post-cutover quality review:

- Split useAIProvider god-module into Zustand aiProviderStore + pure api/aiChat
- Unify AiSidebar/CmdK chat into CopilotChatSession with local transcript state
- Wire ExplorePage onDatasourceChanged into command context (stable callback)
- Drop Vue-shaped analytics router adapter; subscribe to data router directly
- Thin useCommandContext / useRegisterCommandContext to store-native usage
- Decompose >1k files: explore helpers, useLiveLogStream, useTracesExploreData,
  PanelTypeOptions, datasource form sections

All files previously over 1k lines are now under the limit.
