+++ frontend/src/App.tsx
@@ -2,7 +2,6 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { useEffect } from 'react'
 import { RouterProvider } from 'react-router/dom'
 import { initializeAnalytics } from '@/analytics'
-import { createAnalyticsRouterAdapter } from '@/lib/analyticsRouter'
 import { router } from '@/router'
 
 const queryClient = new QueryClient({
@@ -16,7 +15,7 @@ const queryClient = new QueryClient({
 
 export function App() {
   useEffect(() => {
-    void initializeAnalytics(createAnalyticsRouterAdapter(router))
+    void initializeAnalytics(router)
   }, [])
 
   return (
