+++ frontend/src/components/MetricsExplorePanel.tsx
@@ -34,133 +34,28 @@ import {
   type DataSource,
   type DataSourceType,
 } from '@/types/datasource'
-import { dataSourceTypeLogos } from '@/utils/datasourceLogos'
 
-type DatasourceHealthStatus = 'unknown' | 'checking' | 'healthy' | 'unhealthy'
-
-type TraceMetricsNavigationContext = {
-  serviceName?: string
-  startMs?: number
-  endMs?: number
-  createdAt?: number
-}
+import {
+  type DatasourceHealthStatus,
+  type ExploreDatasourceChanged,
+  getTypeLogo,
+  healthLabel,
+  pushQueryHistory,
+  readQueryHistory,
+  TRACE_NAVIGATION_MAX_AGE_MS,
+} from '@/components/explore/exploreShared'
+import {
+  type TraceMetricsNavigationContext,
+  buildServiceMetricsQuery,
+  getDefaultMetricsQuery,
+  getMetricsSmokeQuery,
+  isPrometheusLike,
+  METRICS_HISTORY_KEY,
+  TRACE_METRICS_NAVIGATION_CONTEXT_KEY,
+} from '@/components/explore/metricsExploreHelpers'
 
 type MetricsExplorePanelProps = {
-  onDatasourceChanged?: (payload: { id: string; name: string; type: string }) => void
-}
-
-const HISTORY_KEY = 'explore_query_history'
-const MAX_HISTORY = 10
-const TRACE_METRICS_NAVIGATION_CONTEXT_KEY = 'trace_metrics_navigation'
-const TRACE_NAVIGATION_MAX_AGE_MS = 5 * 60 * 1000
-
-function escapePromQLLabelValue(value: string): string {
-  return value.replace(/\\/g, '\\\\').replace(/"/g, '\\"')
-}
-
-function escapeForSingleQuotedValue(value: string): string {
-  return value.replace(/'/g, "''")
-}
-
-function getDefaultMetricsQuery(type_: DataSourceType): string {
-  switch (type_) {
-    case 'prometheus':
-    case 'victoriametrics':
-      return ''
-    case 'clickhouse':
-      return "SELECT\n  toStartOfInterval(TimeUnix, INTERVAL 1 minute) AS timestamp,\n  avg(Value) AS value,\n  MetricName AS metric\nFROM otel_metrics_gauge\nWHERE TimeUnix >= fromUnixTimestamp({start})\n  AND TimeUnix <= fromUnixTimestamp({end})\nGROUP BY timestamp, metric\nORDER BY timestamp"
-    case 'elasticsearch':
-      return '{"index":"ace-logs","aggs":{"timeseries":{"date_histogram":{"field":"@timestamp","fixed_interval":"1m","min_doc_count":0}}}}'
-    case 'cloudwatch':
-      return '{"namespace":"AWS/EC2","metric_name":"CPUUtilization","stat":"Average","period":60}'
-    default:
-      return ''
-  }
-}
-
-function buildServiceMetricsQuery(type_: DataSourceType, serviceName: string): string {
-  if (type_ === 'clickhouse') {
-    const escapedService = escapeForSingleQuotedValue(serviceName)
-    if (!escapedService) {
-      return 'SELECT timestamp, value, metric FROM metrics WHERE timestamp >= toDateTime({start}) AND timestamp <= toDateTime({end}) ORDER BY timestamp'
-    }
-
-    return `SELECT timestamp, value, metric\nFROM metrics\nWHERE timestamp >= toDateTime({start}) AND timestamp <= toDateTime({end})\nAND service_name = '${escapedService}'\nORDER BY timestamp`
-  }
-
-  if (type_ === 'cloudwatch') {
-    return JSON.stringify(
-      {
-        namespace: 'AWS/ECS',
-        metric_name: 'CPUUtilization',
-        dimensions: serviceName ? { ServiceName: serviceName } : {},
-        stat: 'Average',
-        period: 60,
-      },
-      null,
-      2,
-    )
-  }
-
-  if (type_ === 'elasticsearch') {
-    const serviceFilter = serviceName ? [{ term: { 'service.name.keyword': serviceName } }] : []
-
-    return JSON.stringify(
-      {
-        index: 'ace-logs',
-        query: {
-          bool: {
-            filter: serviceFilter,
-          },
-        },
-        aggs: {
-          timeseries: {
-            date_histogram: {
-              field: '@timestamp',
-              fixed_interval: '30s',
-              min_doc_count: 0,
-            },
-          },
-        },
-      },
-      null,
-      2,
-    )
-  }
-
-  if (!serviceName) {
-    return 'sum(rate(http_requests_total[5m]))'
-  }
-
-  const escapedService = escapePromQLLabelValue(serviceName)
-  return `sum(rate(http_requests_total{service="${escapedService}"}[5m])) or sum(rate(http_requests_total{service_name="${escapedService}"}[5m]))`
-}
-
-function getSmokeQuery(type_: DataSourceType): string {
-  if (type_ === 'prometheus' || type_ === 'victoriametrics') {
-    return 'up'
-  }
-  if (type_ === 'clickhouse') {
-    return "SELECT now() AS timestamp, toFloat64(1) AS value, 'up' AS metric LIMIT 1"
-  }
-  if (type_ === 'cloudwatch') {
-    return '{"namespace":"AWS/EC2","metric_name":"CPUUtilization","stat":"Average","period":60}'
-  }
-  if (type_ === 'elasticsearch') {
-    return '{"index":"ace-logs","aggs":{"timeseries":{"date_histogram":{"field":"@timestamp","fixed_interval":"1m","min_doc_count":0}}}}'
-  }
-  if (type_ === 'loki') {
-    return '{job=~".+"}'
-  }
-  return '*'
-}
-
-function getTypeLogo(type_: DataSourceType): string | undefined {
-  return dataSourceTypeLogos[type_]
-}
-
-function isPrometheusLike(type_: DataSourceType): boolean {
-  return type_ === 'prometheus' || type_ === 'victoriametrics'
+  onDatasourceChanged?: (payload: ExploreDatasourceChanged) => void
 }
 
 export function MetricsExplorePanel({ onDatasourceChanged }: MetricsExplorePanelProps) {
@@ -208,23 +103,10 @@ export function MetricsExplorePanel({ onDatasourceChanged }: MetricsExplorePanel
     ? datasourceHealth[activeDatasource.id] || 'unknown'
     : 'unknown'
 
-  const activeDatasourceHealthLabel =
-    activeDatasourceHealth === 'healthy'
-      ? 'Healthy'
-      : activeDatasourceHealth === 'unhealthy'
-        ? 'Unhealthy'
-        : activeDatasourceHealth === 'checking'
-          ? 'Checking...'
-          : 'Unknown'
+  const activeDatasourceHealthLabel = healthLabel(activeDatasourceHealth)
 
   const addToHistory = useCallback((q: string) => {
-    if (!q.trim()) return
-    setQueryHistory(prev => {
-      const filtered = prev.filter(item => item !== q)
-      const next = [q, ...filtered].slice(0, MAX_HISTORY)
-      sessionStorage.setItem(HISTORY_KEY, JSON.stringify(next))
-      return next
-    })
+    setQueryHistory(prev => pushQueryHistory(METRICS_HISTORY_KEY, prev, q))
   }, [])
 
   const runQuery = useCallback(async () => {
@@ -322,7 +204,7 @@ export function MetricsExplorePanel({ onDatasourceChanged }: MetricsExplorePanel
 
     try {
       const healthResult = await queryDataSource(datasourceId, {
-        query: getSmokeQuery(type_),
+        query: getMetricsSmokeQuery(type_),
         signal:
           type_ === 'clickhouse' || type_ === 'cloudwatch' || type_ === 'elasticsearch'
             ? 'metrics'
@@ -387,14 +269,7 @@ export function MetricsExplorePanel({ onDatasourceChanged }: MetricsExplorePanel
       // Ignore malformed navigation context.
     }
 
-    const stored = sessionStorage.getItem(HISTORY_KEY)
-    if (stored) {
-      try {
-        setQueryHistory(JSON.parse(stored))
-      } catch {
-        setQueryHistory([])
-      }
-    }
+    setQueryHistory(readQueryHistory(METRICS_HISTORY_KEY))
   }, [searchParams])
 
   const previousOrgIdRef = useRef<string | null>(null)
@@ -530,7 +405,7 @@ export function MetricsExplorePanel({ onDatasourceChanged }: MetricsExplorePanel
 
   function clearHistory() {
     setQueryHistory([])
-    sessionStorage.removeItem(HISTORY_KEY)
+    sessionStorage.removeItem(METRICS_HISTORY_KEY)
   }
 
   function toggleDatasourceMenu() {
