+++ frontend/src/components/TracesExplorePanel.tsx
@@ -10,58 +10,34 @@ import {
   Star,
   Waypoints,
 } from 'lucide-react'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useEffect, useRef } from 'react'
 import { useNavigate, useSearchParams } from 'react-router'
-import {
-  fetchDataSourceTrace,
-  fetchDataSourceTraceServiceGraph,
-  fetchDataSourceTraceServices,
-  queryDataSource,
-  searchDataSourceTraces,
-} from '@/api/datasources'
 import { ClickHouseSQLEditor } from '@/components/ClickHouseSQLEditor'
+import {
+  type ExploreDatasourceChanged,
+  CLICKHOUSE_DEFAULT_QUERY,
+  TRACE_NAVIGATION_CONTEXT_KEY,
+  TRACE_NAVIGATION_MAX_AGE_MS,
+  getTypeLogo,
+  healthLabel,
+  writeServiceMetricsNavigationContext,
+  writeTraceLogsNavigationContext,
+} from '@/components/explore/tracesExploreHelpers'
 import { TimeRangePicker } from '@/components/TimeRangePicker'
 import { TraceListPanel } from '@/components/TraceListPanel'
 import { TraceServiceGraph } from '@/components/TraceServiceGraph'
 import { TraceSpanDetailsPanel } from '@/components/TraceSpanDetailsPanel'
 import { TraceTimeline } from '@/components/TraceTimeline'
 import { useTimeRange } from '@/hooks/useTimeRange'
+import { useTracesExploreData } from '@/hooks/useTracesExploreData'
 import { useTracingDatasources } from '@/hooks/useTracingDatasources'
 import { useFavoritesStore } from '@/stores/favoritesStore'
 import { useOrgStore } from '@/stores/orgStore'
-import {
-  dataSourceTypeLabels,
-  type DataSourceType,
-  type Trace,
-  type TraceServiceGraph as TraceServiceGraphModel,
-  type TraceSpan,
-  type TraceSummary,
-} from '@/types/datasource'
-import { convertClickHouseSpansToTraceSummaries } from '@/utils/traceClickHouse'
-import { dataSourceTypeLogos } from '@/utils/datasourceLogos'
+import { dataSourceTypeLabels } from '@/types/datasource'
 import { formatDurationNano, formatTraceStart } from '@/utils/traceFormat'
 
-type DatasourceHealthStatus = 'unknown' | 'checking' | 'healthy' | 'unhealthy'
-
 type TracesExplorePanelProps = {
-  onDatasourceChanged?: (payload: { id: string; name: string; type: string }) => void
-}
-
-const CLICKHOUSE_DEFAULT_QUERY =
-  "SELECT\n  SpanId AS span_id,\n  ParentSpanId AS parent_span_id,\n  SpanName AS operation_name,\n  ServiceName AS service_name,\n  toUnixTimestamp64Nano(Timestamp) AS start_time_unix_nano,\n  Duration AS duration_nano,\n  StatusCode AS status\nFROM ace_traces\nWHERE Timestamp BETWEEN fromUnixTimestamp64Nano({start_ns}) AND fromUnixTimestamp64Nano({end_ns})\nLIMIT 200"
-
-const TRACE_NAVIGATION_CONTEXT_KEY = 'dashboard_trace_navigation'
-const TRACE_LOGS_NAVIGATION_CONTEXT_KEY = 'trace_logs_navigation'
-const TRACE_METRICS_NAVIGATION_CONTEXT_KEY = 'trace_metrics_navigation'
-const TRACE_NAVIGATION_MAX_AGE_MS = 5 * 60 * 1000
-const TRACE_TO_X_PADDING_MS = 5 * 60 * 1000
-
-function getTypeLogo(type_: DataSourceType): string | undefined {
-  return dataSourceTypeLogos[type_]
-}
-
-function toMilliseconds(unixNanoTimestamp: number): number {
-  return Math.floor(unixNanoTimestamp / 1_000_000)
+  onDatasourceChanged?: (payload: ExploreDatasourceChanged) => void
 }
 
 export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelProps) {
@@ -73,345 +49,64 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
   const toggleFavorite = useFavoritesStore(state => state.toggleFavorite)
   const isFavorite = useFavoritesStore(state => state.isFavorite)
 
-  const [selectedDatasourceId, setSelectedDatasourceId] = useState('')
-  const [showDatasourceMenu, setShowDatasourceMenu] = useState(false)
-  const [datasourceHealth, setDatasourceHealth] = useState<Record<string, DatasourceHealthStatus>>({})
-  const [datasourceHealthErrors, setDatasourceHealthErrors] = useState<Record<string, string>>({})
-
-  const [query, setQuery] = useState('')
-  const [selectedService, setSelectedService] = useState('')
-  const [limit, setLimit] = useState(20)
-  const [traceIdInput, setTraceIdInput] = useState('')
-
-  const [loadingSearch, setLoadingSearch] = useState(false)
-  const [loadingTrace, setLoadingTrace] = useState(false)
-  const [loadingServices, setLoadingServices] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [serviceGraphError, setServiceGraphError] = useState<string | null>(null)
-
-  const [services, setServices] = useState<string[]>([])
-  const [traceSummaries, setTraceSummaries] = useState<TraceSummary[]>([])
-  const [selectedTraceId, setSelectedTraceId] = useState('')
-  const [activeTrace, setActiveTrace] = useState<Trace | null>(null)
-  const [activeServiceGraph, setActiveServiceGraph] = useState<TraceServiceGraphModel | null>(null)
-  const [selectedSpan, setSelectedSpan] = useState<TraceSpan | null>(null)
-  const [loadingServiceGraph, setLoadingServiceGraph] = useState(false)
-  const [hasSearched, setHasSearched] = useState(false)
+  const {
+    selectedDatasourceId,
+    setSelectedDatasourceId,
+    showDatasourceMenu,
+    setShowDatasourceMenu,
+    datasourceHealth,
+    datasourceHealthErrors,
+    query,
+    setQuery,
+    selectedService,
+    setSelectedService,
+    limit,
+    setLimit,
+    traceIdInput,
+    setTraceIdInput,
+    loadingSearch,
+    loadingTrace,
+    loadingServices,
+    error,
+    setError,
+    serviceGraphError,
+    services,
+    traceSummaries,
+    selectedTraceId,
+    activeTrace,
+    activeServiceGraph,
+    selectedSpan,
+    setSelectedSpan,
+    loadingServiceGraph,
+    hasSearched,
+    pendingTraceDatasourceIdRef,
+    pendingTraceIdRef,
+    activeDatasource,
+    isClickHouseDatasource,
+    hasTracingDatasources,
+    checkDatasourceHealth,
+    runSearch,
+    loadTrace,
+    loadServices,
+    tryLoadPendingTrace,
+    selectDatasource,
+    resetResults,
+    resetForOrgChange,
+  } = useTracesExploreData({
+    tracingDatasources,
+    timeRange,
+    isCustomRange,
+  })
 
   const datasourceMenuRef = useRef<HTMLDivElement | null>(null)
-  const pendingTraceDatasourceIdRef = useRef('')
-  const pendingTraceIdRef = useRef('')
-
-  const hasTracingDatasources = tracingDatasources.length > 0
-  const activeDatasource = useMemo(
-    () => tracingDatasources.find(ds => ds.id === selectedDatasourceId) ?? null,
-    [tracingDatasources, selectedDatasourceId],
-  )
-  const isClickHouseDatasource = activeDatasource?.type === 'clickhouse'
 
   const activeDatasourceHealth = activeDatasource
     ? datasourceHealth[activeDatasource.id] || 'unknown'
     : 'unknown'
-
-  const activeDatasourceHealthLabel =
-    activeDatasourceHealth === 'healthy'
-      ? 'Healthy'
-      : activeDatasourceHealth === 'unhealthy'
-        ? 'Unhealthy'
-        : activeDatasourceHealth === 'checking'
-          ? 'Checking...'
-          : 'Unknown'
-
+  const activeDatasourceHealthLabel = healthLabel(activeDatasourceHealth)
   const activeDatasourceHealthError = activeDatasource
     ? datasourceHealthErrors[activeDatasource.id] || ''
     : ''
-
-  const checkDatasourceHealth = useCallback(async (datasourceId: string, type_: DataSourceType) => {
-    setDatasourceHealth(prev => ({ ...prev, [datasourceId]: 'checking' }))
-    setDatasourceHealthErrors(prev => {
-      const next = { ...prev }
-      delete next[datasourceId]
-      return next
-    })
-
-    try {
-      if (type_ === 'clickhouse') {
-        const end = Math.floor(Date.now() / 1000)
-        const start = end - 15 * 60
-        const healthResult = await queryDataSource(datasourceId, {
-          query: "SELECT now() AS timestamp, 'up' AS status LIMIT 1",
-          signal: 'traces',
-          start,
-          end,
-          step: 15,
-          limit: 1,
-        })
-
-        if (healthResult.status === 'error') {
-          throw new Error(healthResult.error || 'Health check failed')
-        }
-      } else {
-        await fetchDataSourceTraceServices(datasourceId)
-      }
-
-      setDatasourceHealth(prev => ({ ...prev, [datasourceId]: 'healthy' }))
-    } catch (e) {
-      setDatasourceHealth(prev => ({ ...prev, [datasourceId]: 'unhealthy' }))
-      setDatasourceHealthErrors(prev => ({
-        ...prev,
-        [datasourceId]: e instanceof Error ? e.message : 'Health check failed',
-      }))
-    }
-  }, [])
-
-  const runClickHouseTraceQuery = useCallback(async (overrides?: {
-    datasourceId?: string
-    queryText?: string
-  }) => {
-    // Allow callers that just updated state (e.g. first ClickHouse datasource
-    // selection) to pass fresh values so we do not read a stale closure.
-    const effectiveQuery = (overrides?.queryText ?? query).trim()
-    const effectiveDatasourceId = overrides?.datasourceId ?? selectedDatasourceId
-
-    if (!effectiveQuery) {
-      setError('Query is required')
-      return
-    }
-
-    if (!effectiveDatasourceId) {
-      setError('Select a tracing datasource')
-      return
-    }
-
-    setHasSearched(true)
-    setLoadingSearch(true)
-    setError(null)
-    setActiveTrace(null)
-    setActiveServiceGraph(null)
-    setServiceGraphError(null)
-    setSelectedTraceId('')
-    setSelectedSpan(null)
-
-    try {
-      const start = Math.floor(timeRange.start / 1000)
-      const end = Math.floor(timeRange.end / 1000)
-
-      const response = await queryDataSource(effectiveDatasourceId, {
-        query: effectiveQuery,
-        signal: 'traces',
-        start,
-        end,
-        step: 15,
-        limit,
-      })
-
-      if (response.status === 'error') {
-        setError(response.error || 'Query failed')
-        setTraceSummaries([])
-        return
-      }
-
-      if (response.resultType !== 'traces') {
-        setError('Selected datasource did not return trace results')
-        setTraceSummaries([])
-        return
-      }
-
-      setTraceSummaries(convertClickHouseSpansToTraceSummaries(response.data?.traces || []))
-    } catch (e) {
-      setError(e instanceof Error ? e.message : 'Failed to query traces')
-      setTraceSummaries([])
-    } finally {
-      setLoadingSearch(false)
-    }
-  }, [limit, query, selectedDatasourceId, timeRange.end, timeRange.start])
-
-  const runSearch = useCallback(async () => {
-    if (!selectedDatasourceId) {
-      setError('Select a tracing datasource')
-      return
-    }
-
-    if (isClickHouseDatasource) {
-      await runClickHouseTraceQuery()
-      return
-    }
-
-    setHasSearched(true)
-    setLoadingSearch(true)
-    setError(null)
-
-    try {
-      let start: number
-      let end: number
-
-      if (isCustomRange) {
-        start = Math.floor(timeRange.start / 1000)
-        end = Math.floor(timeRange.end / 1000)
-      } else {
-        const windowDurationSeconds = Math.max(
-          1,
-          Math.floor((timeRange.end - timeRange.start) / 1000),
-        )
-        end = Math.floor(Date.now() / 1000)
-        start = end - windowDurationSeconds
-      }
-
-      const summaries = await searchDataSourceTraces(selectedDatasourceId, {
-        query: query.trim() || undefined,
-        service: selectedService || undefined,
-        start,
-        end,
-        limit,
-      })
-      setTraceSummaries(summaries)
-    } catch (e) {
-      setError(e instanceof Error ? e.message : 'Failed to search traces')
-      setTraceSummaries([])
-    } finally {
-      setLoadingSearch(false)
-    }
-  }, [
-    isClickHouseDatasource,
-    isCustomRange,
-    limit,
-    query,
-    runClickHouseTraceQuery,
-    selectedDatasourceId,
-    selectedService,
-    timeRange.end,
-    timeRange.start,
-  ])
-
-  const loadTrace = useCallback(
-    async (traceId: string) => {
-      if (isClickHouseDatasource) {
-        setError('Trace detail lookup is not available for ClickHouse SQL results yet')
-        return
-      }
-
-      if (!selectedDatasourceId) {
-        setError('Select a tracing datasource')
-        return
-      }
-
-      setLoadingTrace(true)
-      setError(null)
-
-      try {
-        const trace = await fetchDataSourceTrace(selectedDatasourceId, traceId)
-        setActiveTrace(trace)
-        setLoadingServiceGraph(true)
-        setServiceGraphError(null)
-
-        try {
-          const graph = await fetchDataSourceTraceServiceGraph(selectedDatasourceId, traceId)
-          setActiveServiceGraph(graph)
-        } catch (graphError) {
-          setActiveServiceGraph(null)
-          setServiceGraphError(
-            graphError instanceof Error ? graphError.message : 'Failed to fetch trace service graph',
-          )
-        } finally {
-          setLoadingServiceGraph(false)
-        }
-
-        setSelectedTraceId(traceId)
-        setSelectedSpan(null)
-        setTraceIdInput(traceId)
-      } catch (e) {
-        setError(e instanceof Error ? e.message : 'Failed to fetch trace')
-        setActiveTrace(null)
-        setActiveServiceGraph(null)
-        setServiceGraphError(null)
-        setLoadingServiceGraph(false)
-      } finally {
-        setLoadingTrace(false)
-      }
-    },
-    [isClickHouseDatasource, selectedDatasourceId],
-  )
-
-  const loadServices = useCallback(
-    async (autoSearch = false) => {
-      if (isClickHouseDatasource) {
-        setServices([])
-        setSelectedService('')
-        return
-      }
-
-      if (!selectedDatasourceId) {
-        setServices([])
-        setSelectedService('')
-        return
-      }
-
-      setLoadingServices(true)
-      try {
-        const serviceList = await fetchDataSourceTraceServices(selectedDatasourceId)
-        setServices(serviceList)
-        setSelectedService(current =>
-          current && !serviceList.includes(current) ? '' : current,
-        )
-        if (autoSearch) {
-          void runSearch()
-        }
-      } catch {
-        setServices([])
-      } finally {
-        setLoadingServices(false)
-      }
-    },
-    [isClickHouseDatasource, runSearch, selectedDatasourceId],
-  )
-
-  const tryLoadPendingTrace = useCallback(async () => {
-    if (!pendingTraceIdRef.current || !selectedDatasourceId) {
-      return
-    }
-
-    if (
-      pendingTraceDatasourceIdRef.current &&
-      pendingTraceDatasourceIdRef.current !== selectedDatasourceId
-    ) {
-      return
-    }
-
-    const traceId = pendingTraceIdRef.current
-    pendingTraceIdRef.current = ''
-    pendingTraceDatasourceIdRef.current = ''
-    setTraceIdInput(traceId)
-    await loadTrace(traceId)
-  }, [loadTrace, selectedDatasourceId])
-
-  const selectDatasource = useCallback(
-    (datasourceId: string) => {
-      setSelectedDatasourceId(datasourceId)
-      setShowDatasourceMenu(false)
-
-      const ds = tracingDatasources.find(d => d.id === datasourceId)
-      if (ds?.type === 'clickhouse' && !query.trim()) {
-        setQuery(CLICKHOUSE_DEFAULT_QUERY)
-        // Pass fresh values — setState above is not visible to runSearch yet.
-        void runClickHouseTraceQuery({
-          datasourceId,
-          queryText: CLICKHOUSE_DEFAULT_QUERY,
-        })
-      }
-    },
-    [query, runClickHouseTraceQuery, tracingDatasources],
-  )
-
-  const buildNavigationWindow = useCallback(
-    (payload: { startTimeUnixNano: number; endTimeUnixNano: number }) => {
-      const startMs = toMilliseconds(payload.startTimeUnixNano)
-      const endMs = toMilliseconds(payload.endTimeUnixNano)
-      const paddedStartMs = Math.max(0, startMs - TRACE_TO_X_PADDING_MS)
-      const paddedEndMs = Math.max(paddedStartMs + 1_000, endMs + TRACE_TO_X_PADDING_MS)
-      return { startMs: paddedStartMs, endMs: paddedEndMs }
-    },
-    [],
-  )
 
   const openTraceLogs = useCallback(
     (payload: {
@@ -420,24 +115,10 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
       startTimeUnixNano: number
       endTimeUnixNano: number
     }) => {
-      const { startMs, endMs } = buildNavigationWindow(payload)
-      try {
-        localStorage.setItem(
-          TRACE_LOGS_NAVIGATION_CONTEXT_KEY,
-          JSON.stringify({
-            traceId: payload.traceId,
-            serviceName: payload.serviceName || undefined,
-            startMs,
-            endMs,
-            createdAt: Date.now(),
-          }),
-        )
-      } catch {
-        // Ignore localStorage write issues.
-      }
+      writeTraceLogsNavigationContext(payload)
       navigate('/app/explore/logs')
     },
-    [buildNavigationWindow, navigate],
+    [navigate],
   )
 
   const openServiceMetrics = useCallback(
@@ -446,23 +127,10 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
       startTimeUnixNano: number
       endTimeUnixNano: number
     }) => {
-      const { startMs, endMs } = buildNavigationWindow(payload)
-      try {
-        localStorage.setItem(
-          TRACE_METRICS_NAVIGATION_CONTEXT_KEY,
-          JSON.stringify({
-            serviceName: payload.serviceName || undefined,
-            startMs,
-            endMs,
-            createdAt: Date.now(),
-          }),
-        )
-      } catch {
-        // Ignore localStorage write issues.
-      }
+      writeServiceMetricsNavigationContext(payload)
       navigate('/app/explore/metrics')
     },
-    [buildNavigationWindow, navigate],
+    [navigate],
   )
 
   useEffect(() => {
@@ -508,7 +176,7 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
     } catch {
       // Ignore malformed navigation context.
     }
-  }, [searchParams])
+  }, [pendingTraceDatasourceIdRef, pendingTraceIdRef, searchParams, setQuery, setTraceIdInput])
 
   useEffect(() => {
     if (tracingDatasources.length === 0) {
@@ -516,8 +184,6 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
       return
     }
 
-    // Dashboard (and deep-link) navigation always wins over a previously selected
-    // explore datasource so tryLoadPendingTrace is not stuck on a mismatch.
     const pendingDatasourceId = pendingTraceDatasourceIdRef.current
     if (pendingDatasourceId) {
       const pendingDatasource = tracingDatasources.find(ds => ds.id === pendingDatasourceId)
@@ -538,7 +204,14 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
         setQuery(CLICKHOUSE_DEFAULT_QUERY)
       }
     }
-  }, [query, selectedDatasourceId, tracingDatasources])
+  }, [
+    pendingTraceDatasourceIdRef,
+    query,
+    selectedDatasourceId,
+    setQuery,
+    setSelectedDatasourceId,
+    tracingDatasources,
+  ])
 
   useEffect(() => {
     if (!activeDatasource) return
@@ -548,20 +221,12 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
   }, [activeDatasource, checkDatasourceHealth, datasourceHealth])
 
   useEffect(() => {
-    setTraceSummaries([])
-    setActiveTrace(null)
-    setActiveServiceGraph(null)
-    setServiceGraphError(null)
-    setLoadingServiceGraph(false)
-    setSelectedTraceId('')
-    setSelectedSpan(null)
-    setHasSearched(false)
-
+    resetResults()
     void loadServices(true)
     if (!isClickHouseDatasource) {
       void tryLoadPendingTrace()
     }
-  }, [isClickHouseDatasource, loadServices, tryLoadPendingTrace])
+  }, [isClickHouseDatasource, loadServices, resetResults, tryLoadPendingTrace])
 
   useEffect(() => {
     const ds = tracingDatasources.find(d => d.id === selectedDatasourceId)
@@ -580,7 +245,7 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
 
     document.addEventListener('click', handleDocumentClick)
     return () => document.removeEventListener('click', handleDocumentClick)
-  }, [])
+  }, [setShowDatasourceMenu])
 
   useEffect(() => {
     const unsubscribe = onRefresh(() => {
@@ -591,15 +256,10 @@ export function TracesExplorePanel({ onDatasourceChanged }: TracesExplorePanelPr
     return unsubscribe
   }, [hasSearched, loadingSearch, loadingTrace, onRefresh, runSearch, selectedDatasourceId])
 
-  // Reset explore state when org changes
   useEffect(() => {
     void currentOrgId
-    setSelectedDatasourceId('')
-    setDatasourceHealth({})
-    setTraceSummaries([])
-    setActiveTrace(null)
-    setError(null)
-  }, [currentOrgId])
+    resetForOrgChange()
+  }, [currentOrgId, resetForOrgChange])
 
   function toggleDatasourceMenu() {
     if (!hasTracingDatasources || loadingSearch || loadingTrace) {
