+++ frontend/src/hooks/useCommandContext.ts
@@ -1,4 +1,3 @@
-import { useCallback } from 'react'
 import {
   type CommandContext,
   useCommandContextStore,
@@ -6,26 +5,7 @@ import {
 
 export type { CommandContext }
 
-/** Register/deregister page-level context for Cmd+K and the AI sidebar. */
+/** Read the current page-level command context (Cmd+K / AI sidebar). */
 export function useCommandContext() {
-  const currentContext = useCommandContextStore(state => state.currentContext)
-  const register = useCommandContextStore(state => state.registerContext)
-  const deregister = useCommandContextStore(state => state.deregisterContext)
-
-  const registerContext = useCallback(
-    (ctx: CommandContext) => {
-      register(ctx)
-    },
-    [register],
-  )
-
-  const deregisterContext = useCallback(() => {
-    deregister()
-  }, [deregister])
-
-  return {
-    currentContext,
-    registerContext,
-    deregisterContext,
-  }
+  return useCommandContextStore(state => state.currentContext)
 }
