+++ frontend/src/hooks/useRegisterCommandContext.ts
@@ -1,12 +1,13 @@
 import { useEffect } from 'react'
 import {
   type CommandContext,
-  useCommandContext,
-} from '@/hooks/useCommandContext'
+  useCommandContextStore,
+} from '@/stores/commandContextStore'
 
 /** Register page-level Cmd+K / AI sidebar context for the lifetime of the page. */
 export function useRegisterCommandContext(ctx: CommandContext) {
-  const { registerContext, deregisterContext } = useCommandContext()
+  const registerContext = useCommandContextStore(state => state.registerContext)
+  const deregisterContext = useCommandContextStore(state => state.deregisterContext)
 
   // biome-ignore lint/correctness/useExhaustiveDependencies: key on stable context fields, not object identity
   useEffect(() => {
