+++ frontend/src/pages/ExplorePage.tsx
@@ -1,4 +1,4 @@
-import { useMemo } from 'react'
+import { useCallback, useMemo, useState } from 'react'
 import { useNavigate, useParams } from 'react-router'
 import { LogsExplorePanel } from '@/components/LogsExplorePanel'
 import { MetricsExplorePanel } from '@/components/MetricsExplorePanel'
@@ -6,6 +6,12 @@ import { TracesExplorePanel } from '@/components/TracesExplorePanel'
 import { useRegisterCommandContext } from '@/hooks/useRegisterCommandContext'
 
 type ExploreType = 'metrics' | 'logs' | 'traces'
+
+type ExploreDatasourceContext = {
+  id: string
+  name: string
+  type: string
+}
 
 const tabs: { key: ExploreType; label: string }[] = [
   { key: 'metrics', label: 'Metrics' },
@@ -22,17 +28,43 @@ export function ExplorePage() {
   const navigate = useNavigate()
   const { type } = useParams<{ type: string }>()
   const activeType = useMemo(() => normalizeExploreType(type), [type])
+  const [datasource, setDatasource] = useState<ExploreDatasourceContext | null>(null)
 
   useRegisterCommandContext({
     viewName: `Explore · ${activeType[0]!.toUpperCase()}${activeType.slice(1)}`,
     viewRoute: `/app/explore/${activeType}`,
     description: `Explore ${activeType} from connected datasources.`,
+    ...(datasource
+      ? {
+          datasourceId: datasource.id,
+          datasourceType: datasource.type,
+          datasourceName: datasource.name,
+        }
+      : {}),
   })
 
-  function navigateToTab(nextType: ExploreType) {
-    if (nextType === activeType) return
-    navigate(`/app/explore/${nextType}`)
-  }
+  const navigateToTab = useCallback(
+    (nextType: ExploreType) => {
+      if (nextType === activeType) return
+      setDatasource(null)
+      navigate(`/app/explore/${nextType}`)
+    },
+    [activeType, navigate],
+  )
+
+  const handleDatasourceChanged = useCallback((payload: ExploreDatasourceContext) => {
+    setDatasource(current => {
+      if (
+        current &&
+        current.id === payload.id &&
+        current.name === payload.name &&
+        current.type === payload.type
+      ) {
+        return current
+      }
+      return payload
+    })
+  }, [])
 
   return (
     <div className="flex min-w-0 flex-1 flex-col px-8 py-6">
@@ -74,11 +106,11 @@ export function ExplorePage() {
       </nav>
 
       {activeType === 'metrics' ? (
-        <MetricsExplorePanel key="metrics" />
+        <MetricsExplorePanel key="metrics" onDatasourceChanged={handleDatasourceChanged} />
       ) : activeType === 'logs' ? (
-        <LogsExplorePanel key="logs" />
+        <LogsExplorePanel key="logs" onDatasourceChanged={handleDatasourceChanged} />
       ) : (
-        <TracesExplorePanel key="traces" />
+        <TracesExplorePanel key="traces" onDatasourceChanged={handleDatasourceChanged} />
       )}
     </div>
   )
